requirements.txt
shiny
pandas
python-dotenv
arcgis==1.8.1
numpy==1.24.1